+++ p2p/raintree/peers_manager_test.go
@@ -116,7 +116,8 @@ func BenchmarkAddrBookUpdates(b *testing.B) {
 			require.Equal(b, testCase.numExpectedLevels, int(peersManagerStateView.maxNumLevels))
 
 			for i := 0; i < numAddressessToBeAdded; i++ {
-				newAddr, _ := crypto.GenerateAddress()
+				newAddr, err := crypto.GenerateAddress()
+				require.NoError(b, err)
 				network.AddPeerToAddrBook(&types.NetworkPeer{Address: newAddr})
 			}