+++ runtime/test_artifacts/generator.go
@@ -1,19 +1,25 @@
 package test_artifacts
 
+// TODO: Move `root/shared/test_artifacts` to `root/test/test_artifacts`
+// Cross module imports are okay because this is only used for testing and not business logic
 import (
 	"fmt"
 	"math/big"
 	"strconv"
 
-	typesPersistence "github.com/pokt-network/pocket/persistence/types"
-	"github.com/pokt-network/pocket/shared/modules"
-	"github.com/pokt-network/pocket/utility/types"
-
+	typesCons "github.com/pokt-network/pocket/consensus/types"
+	typesP2P "github.com/pokt-network/pocket/p2p/types"
+	typesPers "github.com/pokt-network/pocket/persistence/types"
+	"github.com/pokt-network/pocket/runtime"
 	"github.com/pokt-network/pocket/shared/crypto"
+	"github.com/pokt-network/pocket/shared/modules"
+	typesTelemetry "github.com/pokt-network/pocket/telemetry"
+	"github.com/pokt-network/pocket/utility/types"
+	typesUtil "github.com/pokt-network/pocket/utility/types"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
-// TODO (Team)/INVESTIGATE(olshansy) It seems improperly scoped that the modules have to have shared 'testing' code
+// INVESTIGATE: It seems improperly scoped that the modules have to have shared 'testing' code
 //  It might be an inevitability to have shared testing code, but would like more eyes on it.
 //  Look for opportunities to make testing completely modular
 
@@ -34,28 +40,31 @@ var (
 
 // TODO (Team) this is meant to be a **temporary** replacement for the recently deprecated
 // 'genesis config' option. We need to implement a real suite soon!
-func NewGenesisState(numValidators, numServiceNodes, numApplications, numFisherman int) (genesisState modules.GenesisState, validatorPrivateKeys []string) {
-	apps, appsPrivateKeys := NewActors(MockActorType_App, numApplications)
-	vals, validatorPrivateKeys := NewActors(MockActorType_Val, numValidators)
-	serviceNodes, snPrivateKeys := NewActors(MockActorType_Node, numServiceNodes)
-	fish, fishPrivateKeys := NewActors(MockActorType_Fish, numFisherman)
-	return modules.GenesisState{
-		ConsensusGenesisState: &MockConsensusGenesisState{
+func NewGenesisState(numValidators, numServiceNodes, numApplications, numFisherman int) (modules.GenesisState, []string) {
+	apps, appsPrivateKeys := NewActors(types.ActorType_App, numApplications)
+	vals, validatorPrivateKeys := NewActors(types.ActorType_Validator, numValidators)
+	serviceNodes, snPrivateKeys := NewActors(types.ActorType_ServiceNode, numServiceNodes)
+	fish, fishPrivateKeys := NewActors(types.ActorType_Fisherman, numFisherman)
+
+	genesisState := runtime.NewGenesis(
+		&typesCons.ConsensusGenesisState{
 			GenesisTime:   timestamppb.Now(),
 			ChainId:       DefaultChainID,
 			MaxBlockBytes: DefaultMaxBlockBytes,
-			Validators:    vals,
+			Validators:    typesCons.ToConsensusValidators(vals),
 		},
-		PersistenceGenesisState: &MockPersistenceGenesisState{
-			Pools:        NewPools(),
-			Accounts:     NewAccounts(numValidators+numServiceNodes+numApplications+numFisherman, append(append(append(validatorPrivateKeys, snPrivateKeys...), fishPrivateKeys...), appsPrivateKeys...)...), // TODO(olshansky): clean this up
-			Applications: apps,
-			Validators:   vals,
-			ServiceNodes: serviceNodes,
-			Fishermen:    fish,
-			Params:       DefaultParams(),
+		&typesPers.PersistenceGenesisState{
+			Pools:        typesPers.ToPersistenceAccounts(NewPools()),
+			Accounts:     typesPers.ToPersistenceAccounts(NewAccounts(numValidators+numServiceNodes+numApplications+numFisherman, append(append(append(validatorPrivateKeys, snPrivateKeys...), fishPrivateKeys...), appsPrivateKeys...)...)), // TODO(olshansky): clean this up
+			Applications: typesPers.ToPersistenceActors(apps),
+			Validators:   typesPers.ToPersistenceActors(vals),
+			ServiceNodes: typesPers.ToPersistenceActors(serviceNodes),
+			Fishermen:    typesPers.ToPersistenceActors(fish),
+			Params:       typesPers.ToPersistenceParams(DefaultParams()),
 		},
-	}, validatorPrivateKeys
+	)
+
+	return genesisState, validatorPrivateKeys
 }
 
 func NewDefaultConfigs(privateKeys []string) (configs []modules.Config) {
@@ -66,53 +75,54 @@ func NewDefaultConfigs(privateKeys []string) (configs []modules.Config) {
 }
 
 func NewDefaultConfig(i int, pk string) modules.Config {
-	return modules.Config{
-		Base: &modules.BaseConfig{
+	return runtime.NewConfig(
+		&runtime.BaseConfig{
 			RootDirectory: "/go/src/github.com/pocket-network",
 			PrivateKey:    pk,
 		},
-		Consensus: &MockConsensusConfig{
-			MaxMempoolBytes: 500000000,
-			PacemakerConfig: &MockPacemakerConfig{
-				TimeoutMsec:               5000,
-				Manual:                    true,
-				DebugTimeBetweenStepsMsec: 1000,
-			},
-			PrivateKey: pk,
-		},
-		Utility: &MockUtilityConfig{
+		runtime.WithConsensusConfig(
+			&typesCons.ConsensusConfig{
+				MaxMempoolBytes: 500000000,
+				PacemakerConfig: &typesCons.PacemakerConfig{
+					TimeoutMsec:               5000,
+					Manual:                    true,
+					DebugTimeBetweenStepsMsec: 1000,
+				},
+				PrivateKey: pk,
+			}),
+		runtime.WithUtilityConfig(&typesUtil.UtilityConfig{
 			MaxMempoolTransactionBytes: 1024 * 1024 * 1024, // 1GB V0 defaults
 			MaxMempoolTransactions:     9000,
-		},
-		Persistence: &typesPersistence.PersistenceConfig{
+		}),
+		runtime.WithPersistenceConfig(&typesPers.PersistenceConfig{
 			PostgresUrl:    "postgres://postgres:postgres@pocket-db:5432/postgres",
 			NodeSchema:     "node" + strconv.Itoa(i+1),
 			BlockStorePath: "/var/blockstore",
-		},
-		P2P: &MockP2PConfig{
+		}),
+		runtime.WithP2PConfig(&typesP2P.P2PConfig{
 			ConsensusPort:         8080,
 			UseRainTree:           true,
 			IsEmptyConnectionType: false,
 			PrivateKey:            pk,
-		},
-		Telemetry: &MockTelemetryConfig{
+		}),
+		runtime.WithTelemetryConfig(&typesTelemetry.TelemetryConfig{
 			Enabled:  true,
 			Address:  "0.0.0.0:9000",
 			Endpoint: "/metrics",
-		},
-	}
+		}),
+	)
 }
 
 func NewPools() (pools []modules.Account) { // TODO (Team) in the real testing suite, we need to populate the pool amounts dependent on the actors
-	for _, name := range typesPersistence.PoolNames_name {
-		if name == typesPersistence.PoolNames_FeeCollector.String() {
-			pools = append(pools, &MockAcc{
+	for _, name := range typesPers.PoolNames_name {
+		if name == typesPers.PoolNames_FeeCollector.String() {
+			pools = append(pools, &typesPers.Account{
 				Address: name,
 				Amount:  "0",
 			})
 			continue
 		}
-		pools = append(pools, &MockAcc{
+		pools = append(pools, &typesPers.Account{
 			Address: name,
 			Amount:  DefaultAccountAmountString,
 		})
@@ -127,7 +137,7 @@ func NewAccounts(n int, privateKeys ...string) (accounts []modules.Account) {
 			pk, _ := crypto.NewPrivateKey(privateKeys[i])
 			addr = pk.Address().String()
 		}
-		accounts = append(accounts, &MockAcc{
+		accounts = append(accounts, &typesPers.Account{
 			Address: addr,
 			Amount:  DefaultAccountAmountString,
 		})
@@ -135,10 +145,10 @@ func NewAccounts(n int, privateKeys ...string) (accounts []modules.Account) {
 	return
 }
 
-func NewActors(actorType MockActorType, n int) (actors []modules.Actor, privateKeys []string) {
+func NewActors(actorType typesUtil.ActorType, n int) (actors []modules.Actor, privateKeys []string) {
 	for i := 0; i < n; i++ {
 		genericParam := fmt.Sprintf("node%d.consensus:8080", i+1)
-		if int32(actorType) == int32(MockActorType_App) {
+		if int32(actorType) == int32(types.ActorType_App) {
 			genericParam = DefaultMaxRelaysString
 		}
 		actor, pk := NewDefaultActor(int32(actorType), genericParam)
@@ -151,12 +161,12 @@ func NewActors(actorType MockActorType, n int) (actors []modules.Actor, privateK
 func NewDefaultActor(actorType int32, genericParam string) (actor modules.Actor, privateKey string) {
 	privKey, pubKey, addr := GenerateNewKeysStrings()
 	chains := DefaultChains
-	if actorType == int32(typesPersistence.ActorType_Val) {
+	if actorType == int32(typesPers.ActorType_Val) {
 		chains = nil
-	} else if actorType == int32(MockActorType_App) {
+	} else if actorType == int32(types.ActorType_App) {
 		genericParam = DefaultMaxRelaysString
 	}
-	return &MockActor{
+	return &typesPers.Actor{
 		Address:         addr,
 		PublicKey:       pubKey,
 		Chains:          chains,
@@ -165,7 +175,7 @@ func NewDefaultActor(actorType int32, genericParam string) (actor modules.Actor,
 		PausedHeight:    DefaultPauseHeight,
 		UnstakingHeight: DefaultUnstakingHeight,
 		Output:          addr,
-		ActorType:       MockActorType(actorType),
+		ActorType:       typesPers.ActorType(actorType),
 	}, privKey
 }