+++ shared/modules/persistence_module.go
@@ -9,6 +9,8 @@ import (
 
 type PersistenceModule interface {
 	Module
+	ConfigurableModule
+	GenesisDependentModule
 
 	// Context interface
 	NewRWContext(height int64) (PersistenceRWContext, error)
@@ -18,6 +20,7 @@ type PersistenceModule interface {
 	// BlockStore interface
 	ResetContext() error // DEPRECATE(#252): Remove in #284 per the interface changes in #252
 	GetBlockStore() kvstore.KVStore
+	NewWriteContext() PersistenceRWContext
 
 	// Debugging / development only
 	HandleDebugMessage(*debug.DebugMessage) error
@@ -47,7 +50,6 @@ type PersistenceWriteContext interface {
 	// Context Operations
 	NewSavePoint([]byte) error
 	RollbackToSavePoint([]byte) error
-
 	Release() error
 
 	// Block / indexer operations